# Conflicts:
#	src/main/java/org/patryk3211/powergrid/collections/ModdedItems.java
#	src/main/java/org/patryk3211/powergrid/ponder/PowerGridPonderTags.java
#	src/main/resources/assets/powergrid/lang/default/components.json

+++ src/main/java/org/patryk3211/powergrid/electricity/base/ThermalBehaviour.java
@@ -302,9 +302,9 @@ public class ThermalBehaviour extends BlockEntityBehaviour implements ISynchroni
                 float chance = (temperature - overheatTemperature + 100) / 100;
                 if (random.nextFloat() < chance) {
                     if (particleGenerator == null) {
-                        float x = pos.getX() + random.nextFloat();
-                        float y = pos.getY() + random.nextFloat();
-                        float z = pos.getZ() + random.nextFloat();
+                        double x = pos.getX() + random.nextDouble();
+                        double y = pos.getY() + random.nextDouble();
+                        double z = pos.getZ() + random.nextDouble();
                         world.addParticle(ParticleTypes.SMOKE, x, y, z, 0.0f, 0.05f, 0.0f);
                     } else {
                         particleGenerator.generate((x, y, z) ->
@@ -415,6 +415,6 @@ public class ThermalBehaviour extends BlockEntityBehaviour implements ISynchroni
 
     @FunctionalInterface
     public interface IParticleConsumer {
-        void accept(float x, float y, float z);
+        void accept(double x, double y, double z);
     }
 }
